# openapi.yml
openapi: 3.0.3
info:
  title: Swagger API de gestão de processos seletivos - OpenAPI 3.0
  description: |-
    Está é uma API Rest que tem como objetivo expor endpoints para realizar a gestão de um processo seletivo.
    
    Essa API Rest faz parte de um trabalho de conclusão de curso de Sistemas de Informação do aluno Lucas Moraes Cavalcante na Universidade Veiga de Almeida.
  termsOfService: https://github.com/LMoraes7/recruitment-selection
  contact:
    name: Lucas Moraes Cavalcante
    email: lukaocavalcante93@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:8080
...
paths:
...
  '/candidacy/close/{candidacyIdentifier}/selective-process/{selectiveProcessIdentifier}':
    patch:
      tags:
        - candidacy
      summary: Solicitar encerramento da candidatura
      description: |-
        Endpoint responsável por encerrar uma candidatura.
        <br /> <br />
        **Scope:** FUNC_CLOSE_CANDIDACY
      operationId: closeCandidacy
      parameters:
        - $ref: '#/components/parameters/CandidacyIdentifierPath'
        - $ref: '#/components/parameters/SelectiveProcessIdentifierPath'
        - $ref: '#/components/parameters/AuthorizationHeader'
      responses:
        '204':
          description: Solicitação efetuada com sucesso.
        '401':
          description: Solicitação não autorizada devido a falta de autenticação.
        '403':
          description: Solicitação não autorizada devido a falta de permissão.
        '404':
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                candidacySelectiveProcessNotFoundErrorResponse:
                  $ref: '#/components/examples/CandidacySelectiveProcessNotFoundErrorResponse'
        '500':
          description: Ocorreu um erro interno no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                internalErrorResponse:
                  $ref: '#/components/examples/InternalErrorResponse'
...
  '/candidacy/selective-process/{selectiveProcessIdentifier}':
    post:
      tags:
        - candidacy
      summary: Solicitar criação de uma candidatura
      description: |-
        Endpoint responsável por criar uma candidatura.
        <br /> <br />
        **Scope:** FUNC_CREATE_CANDIDACY
      operationId: createCandidacy
      parameters:
        - $ref: '#/components/parameters/SelectiveProcessIdentifierPath'
        - $ref: '#/components/parameters/AuthorizationHeader'
      responses:
        '201':
          description: Solicitação efetuada com sucesso.
        '401':
          description: Solicitação não autorizada devido a falta de autenticação.
        '403':
          description: Solicitação não autorizada devido a falta de permissão.
        '404':
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                selectiveProcessNotFoundErrorResponse:
                  $ref: '#/components/examples/SelectiveProcessNotFoundErrorResponse'
        '422':
          description: A requisição não pode ser processada devido a um erro de negócio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                createCandidacyErrorResponse:
                  $ref: '#/components/examples/CreateCandidacyErrorResponse'
        '500':
          description: Ocorreu um erro interno no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                internalErrorResponse:
                  $ref: '#/components/examples/InternalErrorResponse'
  '/selective-process':
    get:
      tags:
        - selective process
      summary: Solicitar busca de processos seletivos em aberto
      description: |-
        Endpoint responsável por buscar processos seletivos em aberto.
      operationId: consultListOfSelectiveProcess
      parameters:
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/PageNumberParam'
      responses:
        '200':
          description: Solicitação efetuada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelectiveProcessDataPaginatedResponse'
              examples:
                selectiveProcessDataPaginatedResponse:
                  $ref: '#/components/examples/SelectiveProcessDataPaginatedResponse'
        '204':
          description: Solicitação efetuada com sucesso, porém não foi encontrada nenhum processo seletivo com os parâmetros de paginação informados.
        '400':
          description: A requisição contém dados inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                badRequestErrorResponse:
                  $ref: '#/components/examples/BadRequestErrorResponse'
        '500':
          description: Ocorreu um erro interno no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                internalErrorResponse:
                  $ref: '#/components/examples/InternalErrorResponse'

    post:
      tags:
        - selective process
      summary: Solicitar criação de um processo seletivo
      description: |-
        Endpoint responsável por criar um processo seletivo.
        <br /> <br />
        **Scope:** FUNC_CREATE_SELECTIVE_PROCESS
      operationId: createSelectiveProcess
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        description: Dados para realizar a criação de um processo seletivo.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSelectiveProcessRequest'
            examples:
              CreateSelectiveProcessRequest:
                $ref: '#/components/examples/CreateSelectiveProcessRequest'
      responses:
        '201':
          description: Solicitação efetuada com sucesso.
        '400':
          description: A requisição contém dados inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                badRequestErrorResponse:
                  $ref: '#/components/examples/BadRequestErrorResponse'
        '401':
          description: Solicitação não autorizada devido a falta de autenticação.
        '403':
          description: Solicitação não autorizada devido a falta de permissão.
        '422':
          description: A requisição não pode ser processada devido a um erro de negócio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                stepInvalidErrorResponse:
                  $ref: '#/components/examples/StepInvalidErrorResponse'
                externalStepInvalidErrorResponse:
                  $ref: '#/components/examples/ExternalStepInvalidErrorResponse'
        '500':
          description: Ocorreu um erro interno no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                internalErrorResponse:
                  $ref: '#/components/examples/InternalErrorResponse'
  '/selective-process/{selectiveProcessIdentifier}':
    get:
      tags:
        - selective process
      summary: Solicitar busca de um processo seletivo específico
      description: |-
        Endpoint responsável por buscar processo seletivo específico.
      operationId: consultSpecificSelectiveProcess
      parameters:
        - $ref: '#/components/parameters/SelectiveProcessIdentifierPath'
      responses:
        '200':
          description: Solicitação efetuada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelectiveProcessResponse'
              examples:
                selectiveProcessResponse:
                  $ref: '#/components/examples/SelectiveProcessResponse'
        '404':
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                selectiveProcessNotFoundErrorResponse:
                  $ref: '#/components/examples/SelectiveProcessNotFoundErrorResponse'
        '500':
          description: Ocorreu um erro interno no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                internalErrorResponse:
                  $ref: '#/components/examples/InternalErrorResponse'

    patch:
      tags:
        - selective process
      summary: Solicitar encerramento de um processo seletivo
      description: |-
        Endpoint responsável por encerrar um processo seletivo.
        <br /> <br />
        **Scope:** FUNC_CLOSE_SELECTIVE_PROCESS
      operationId: closeSelectiveProcess
      parameters:
        - $ref: '#/components/parameters/SelectiveProcessIdentifierPath'
        - $ref: '#/components/parameters/AuthorizationHeader'
      responses:
        '204':
          description: Solicitação efetuada com sucesso.
        '401':
          description: Solicitação não autorizada devido a falta de autenticação.
        '403':
          description: Solicitação não autorizada devido a falta de permissão.
        '404':
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                selectiveProcessNotFoundErrorResponse:
                  $ref: '#/components/examples/SelectiveProcessNotFoundErrorResponse'
        '500':
          description: Ocorreu um erro interno no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultErrorResponse'
              examples:
                internalErrorResponse:
                  $ref: '#/components/examples/InternalErrorResponse'
...
components:
  schemas:
...
    DefaultErrorResponse:
      type: object
      description: Dados informando algum erro retornado para o cliente.
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Campo informando o código de erro que aconteceu.
          nullable: false
        message:
          type: string
          description: Mensagem para dar maior detalhamento do erro ocorrido.
          nullable: false
        errors:
          type: array
          description: Informações sobre os campos que causaram o erro.
          nullable: true
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
                description: Nome do campo que causou o erro gerado.
                nullable: false
              message:
                type: string
                description: Mensagem para dar maior detalhamento do erro ocorrido no campo.
                nullable: false
...
    SelectiveProcessDataPaginatedResponse:
      type: object
      description: Informações sobre os processos seletivos buscados.
      required:
        - selective_processes
        - page_number
        - page_size
        - total_pages
        - total_elements
        - total_results
      properties:
        selective_processes:
          $ref: '#/components/schemas/SelectiveProcessPaginatedResponse'
        page_number:
          type: integer
          format: int64
          description: Número da página utilizado no filtro.
          nullable: false
        page_size:
          type: integer
          format: int64
          description: Quantidade de ítens na página utilizado no filtro.
          nullable: false
        total_pages:
          type: integer
          format: int64
          description: Quantidade total de páginas com o filtro de busca utilizado.
          nullable: false
        total_elements:
          type: integer
          format: int64
          description: Quantidade de ítens páginas que foram retornados.
          nullable: false
        total_results:
          type: integer
          format: int64
          description: Quantidade total de ítens com o filtro de busca utilizado.
          nullable: false
    SelectiveProcessPaginatedResponse:
      type: array
      description: Informações sobre os processos seletivos.
      nullable: false
      items:
        type: object
        required:
          - id
          - title
        properties:
          id:
            type: string
            description: Identificador do processo seletivo.
            nullable: false
          title:
            type: string
            description: Título do processo seletivo.
            nullable: false
    SelectiveProcessResponse:
      type: object
      description: Informações sobre um processo seletivo buscado.
      required:
        - id
        - title
        - description
        - desired_position
        - status
        - responsibilities
        - requirements
      properties:
        id:
          type: string
          description: Identificador do processo seletivo.
          nullable: false
        title:
          type: string
          description: Título do processo seletivo.
          nullable: false
        description:
          type: string
          description: Descrição do processo seletivo
          nullable: false
        desired_position:
          type: string
          description: Cargo desejado para contratação no processo seletivo
          nullable: false
        status:
          type: string
          description: Status do processo seletivo. Em progresso (IN_PROGRESS) e Encerrado (CLOSED).
          enum:
            - IN_PROGRESS
            - CLOSED
          nullable: false
        responsibilities:
          type: array
          description: Responsabilidades descritas para o cargo.
          nullable: false
          items:
            type: string
            nullable: false
        requirements:
          type: array
          description: Requerimentos obrigatórios para o cargo.
          nullable: false
          items:
            type: string
            nullable: false
        additional_infos:
          type: array
          description: Informações adicionais do processo seletivo.
          nullable: true
          items:
            type: string
            nullable: false
    CreateSelectiveProcessRequest:
      type: object
      description: Dados para solicitar a crição de um processo seletivo.
      required:
        - title
        - description
        - desired_position
        - responsibilities
        - requirements
        - steps
      properties:
        title:
          type: string
          description: Título do processo seletivo.
          nullable: false
        description:
          type: string
          description: Descrição do processo seletivo.
          nullable: false
        desired_position:
          type: string
          description: Cargo desejado para contratação no processo seletivo.
          nullable: false
        responsibilities:
          type: array
          description: Responsabilidades descritas para o cargo.
          nullable: false
          items:
            type: string
            nullable: false
        requirements:
          type: array
          description: Requerimentos obrigatórios para o cargo.
          nullable: false
          items:
            type: string
            nullable: false
        additional_infos:
          type: array
          description: Informações adicionais do processo seletivo.
          nullable: true
          items:
            type: string
            nullable: false
        steps:
          $ref: '#/components/schemas/SelectiveProcessStepRequest'
    SelectiveProcessStepRequest:
      type: array
      description: Informações sobre as etapas para compor um processo seletivo.
      minItems: 2
      maxItems: 5
      nullable: false
      items:
        type: object
        required:
          - identifier
          - limit_time
        properties:
          identifier:
            type: string
            description: Identificador da etapa.
            nullable: false
          limit_time:
            type: integer
            format: int64
            description: |-
              Tempo para etapa ser realizada após a sua liberação para o candidato.
              <br />
              Deve-se informar o tempo em dias. 
              <br />
              Caso a etapa seja do externa, não deve preencher esse campo, pois não é permitido definir um tempo limite para etapas externas.
            minimum: 1
            nullable: true
...
  examples:
...
    SelectiveProcessDataPaginatedResponse:
      summary: "Exemplo de resposta para busca de processos seletivos"
      value:
        selective_processes:
          - id: "SEL-KjoiJ77gy"
            title: "Desenvolvedor Backend Java Pleno"
          - id: "SEL-ljUiJQ1gy"
            title: "Desenvolvedor Backend Java Sênior"
          - id: "SEL-Mjop871gy"
            title: "Desenvolvedor Backend Java Júnior"
        page_number: 0
        page_size: 10
        total_pages: 1
        total_elements: 3
        total_results: 3
...
    SelectiveProcessResponse:
      summary: "Exemplo de resposta para busca de um processo seletivo"
      value:
        id: "SEL-KjoiJ77gy"
        title: "Desenvolvedor Backend Java Pleno"
        description: "Processo seletivo para contratação de uma vaga para desenvolvedor Java pleno"
        desired_position: "Desenvolvedor Java Pleno"
        status: "IN_PROGRESS"
        responsibilities: [
          "Realizar manutenção no sistema",
          "Criar novas funcionalidades",
          "Atuar junto ao time de produto"
        ]
        requirements: [
          "Experiência de +3 anos",
          "Conhecimento em Spring Framework",
          "Familiaridade com ambiente AWS"
        ]
        additional_infos: [
          "VR + VA flexível",
          "Atuação remota de trabalho",
          "PLR"
        ]
    CreateSelectiveProcessRequest:
      summary: "Exemplo de solicitação para criação de um processo seletivo"
      value:
        title: "Desenvolvedor Backend Java Pleno"
        description: "Processo seletivo para contratação de uma vaga para desenvolvedor Java pleno"
        desired_position: "Desenvolvedor Java Pleno"
        responsibilities: [
          "Realizar manutenção no sistema",
          "Criar novas funcionalidades",
          "Atuar junto ao time de produto"
        ]
        requirements: [
          "Experiência de +3 anos",
          "Conhecimento em Spring Framework",
          "Familiaridade com ambiente AWS"
        ]
        additional_infos: [
          "VR + VA flexível",
          "Atuação remota de trabalho",
          "PLR"
        ]
        steps:
          - identifier: "STE-Nuy7H56KJ"
            limit_time: 2
          - identifier: "STE-oNKYfc65H"
          - identifier: "STE-Kamknsd8d"
            limit_time: 7
...
    BadRequestErrorResponse:
      summary: "Exemplo genérico de resposta de erro na requisição do cliente"
      value:
        code: "INVALID_ARGUMENTS"
        message: "Informações erradas foram fornecidas. Por favor, revise o contrato."
        errors:
          - field: "abc.xpto"
            message: "não deve estar em branco"
          - field: "xyz"
            message: "deve corresponder a \"ABC|DEF\""
          - field: "abc.dto"
            message: "não pode ser nulo"
...
    InternalErrorResponse:
      summary: "Exemplo de resposta de erro interno no servidor"
      value:
        code: "INTERNAL_ERROR"
        message: "Aconteceu um erro inesperado. Caso o erro persista por um tempo, por favor, entre em contato com o administrador do sistema."
...
    CandidacySelectiveProcessNotFoundErrorResponse:
      summary: "Exemplo de resposta informando que não existe relação entre a candidatura e o processo seletivo informado"
      value:
        code: "NOT_FOUND"
        message: "As entidades de identificadores APP-KJ89dh3G1 e SEL-l1Hp7dJBJ não possuem relação entre si."
...
    SelectiveProcessNotFoundErrorResponse:
      summary: "Exemplo de resposta informando que não existe o processo seletivo informado"
      value:
        code: "NOT_FOUND"
        message: "A entidade de identificador SEL-KJ89dh3G1 não existe."
    CreateCandidacyErrorResponse:
      summary: "Exemplo de resposta informando que não é possível efetuar uma candidatura em um processo seletivo que se encontra fechado"
      value:
        code: "APIX_012"
        message: "Processo seletivo informado se encontra fechado para novas candidaturas."
    StepInvalidErrorResponse:
      summary: "Exemplo de resposta informando que não é possível cadastrar um processo seletivo pois foram informados identificadores de etapas inválidos"
      value:
        code: "APIX_010"
        message: "Os identificadores fornecidos são inválidos: [STE-bhjb675Ga, STE-Ghkb675Ga]."
    ExternalStepInvalidErrorResponse:
      summary: "Exemplo de resposta informando que não é possível cadastrar um processo seletivo pois foram informados tempo limite de execução para etapas externas"
      value:
        code: "APIX_011"
        message: "Os identificadores fornecidos são inválidos: [STE-bhjb675Ga, STE-Ghkb675Ga]."
...
  parameters:
    AuthorizationHeader:
      name: Authorization
      in: header
      description: >-
        Token de acesso no formato JWT.
        
        > Obs.: para mais informações sobre como gerar o token de autenticação, acesso o endpoint '/login'.
      required: true
      schema:
        type: string
    CandidacyIdentifierPath:
      name: candidacyIdentifier
      in: path
      description: Identificador de uma candidatura.
      required: true
      schema:
        type: string
...
    SelectiveProcessIdentifierPath:
      name: selectiveProcessIdentifier
      in: path
      description: Identificador de um processo seletivo.
      required: true
      schema:
        type: string
    PageSizeParam:
      name: pageSize
      in: query
      description: Tamanho de ítens a serem retornados na página.
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 10
    PageNumberParam:
      name: pageNumber
      in: query
      description: Número da página.
      required: true
      schema:
        type: integer
        minimum: 0